requirements.txt
fastapi==0.115.13
loguru==0.7.3
pydantic==2.11.7
PyJWT==2.10.1
pyotp==2.9.0
rtoml==0.12.0
uvicorn==0.34.3
watchdog==6.0.0
